feat & test & feat!: add onCollision override in Melee, add TestMelee test class, light adjustements to TestBaseEnemy test class, remove player field from BaseEnemy, onCollision method in BaseEnemy now has an empty body.
An explanatory comment has been added over the onCollision method in BaseEnemy regarding why it is now empty. The goal is to let future decolators for BaseEnemy take responsibility for the behaviour "onCollision" should have, this is because when an enemy physically collides with a player, the enemy should not necessarily react to such event, it depends on what kind of enemy it is. For example, a Melee enemy will deal damage to the player after an onCollision gets triggered by it.