fix: fix issue regarding gameObjects scaling on the gui.
The images that represent gameObjects were not getting resized on creation to match previous window resizes.
One cause was that, for example, after the window size gets increased, the gameObjects added after that resizement
appeared smaller than what they should be, and vice versa if the window size gets decreased.
To solve this issue I added an extra multiplier in the formula inside the binding between ImageViews and the scene size,
this extra multiplier increases and decreases according to how much the window is resized, this way the new Images
added will be resized correctly.